pyproject.toml
[project]
name = "youtube-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.5",
    "flask>=3.1.3",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "youtube-transcript-api>=1.2.4",
]
